+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/mapper/MulitiQueryRowMapper3Impl.java
@@ -10,16 +10,16 @@ package cn.featherfly.hammer.sqldb.jdbc.mapper;
 
 import java.io.Serializable;
 import java.util.Map;
+import java.util.function.BiFunction;
+import java.util.function.Function;
 
 import org.apache.commons.lang3.ArrayUtils;
 
-import cn.featherfly.common.bean.InstantiatorFactory;
 import cn.featherfly.common.repository.mapper.MulitiQueryRowMapper3;
 import cn.featherfly.common.repository.mapper.MulitiQueryRowMapper4;
+import cn.featherfly.common.repository.mapper.PrefixedBeanMapper;
 import cn.featherfly.common.repository.mapper.RowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.Jdbc;
-import cn.featherfly.hammer.sqldb.jdbc.MapRowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.NestedBeanPropertyRowMapper;
+import cn.featherfly.common.repository.mapper.TupleRowMapperBuilder;
 
 /**
  * MulitiQueryRowMapper3.
@@ -36,36 +36,44 @@ public class MulitiQueryRowMapper3Impl<T1, T2, T3> extends AbstractMulitiQueryRo
      * Instantiates a new muliti query row mapper 3 impl.
      *
      * @param rowMappers the row mappers
-     * @param jdbc the jdbc
-     * @param instantiatorFactory the instantiator factory
+     * @param getRowMapper the get row mapper
      */
-    public MulitiQueryRowMapper3Impl(RowMapper<?>[] rowMappers, Jdbc jdbc, InstantiatorFactory instantiatorFactory) {
-        super(rowMappers, jdbc, instantiatorFactory);
+    public MulitiQueryRowMapper3Impl(RowMapper<?>[] rowMappers,
+        BiFunction<Class<?>, String, RowMapper<?>> getRowMapper) {
+        super(rowMappers, getRowMapper);
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public MulitiQueryRowMapper4<T1, T2, T3, Map<String, Serializable>> map() {
-        return map(new MapRowMapper(jdbc.getSqlTypeMappingManager()));
+    public MulitiQueryRowMapper4<T1, T2, T3, Map<String, Serializable>> mapper() {
+        return mapper(getRowMapper(PrefixedBeanMapper.MAP_TYPE));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T4> MulitiQueryRowMapper4<T1, T2, T3, T4> map(Class<T4> mappingType) {
-        return map(new NestedBeanPropertyRowMapper<>(instantiatorFactory.create(mappingType),
-            jdbc.getSqlTypeMappingManager()));
+    public <T4> MulitiQueryRowMapper4<T1, T2, T3, T4> mapper(Class<T4> mappingType) {
+        return mapper(getRowMapper(mappingType));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T4> MulitiQueryRowMapper4<T1, T2, T3, T4> map(RowMapper<T4> rowMapper) {
-        return new MulitiQueryRowMapper4Impl<>(ArrayUtils.add(getRowMappers(), rowMapper), jdbc, instantiatorFactory);
+    public <T4> MulitiQueryRowMapper4<T1, T2, T3, T4> mapper(RowMapper<T4> rowMapper) {
+        return new MulitiQueryRowMapper4Impl<>(ArrayUtils.add(getRowMappers(), rowMapper),
+            getRowMapper);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public <T4> MulitiQueryRowMapper4<T1, T2, T3, T4> mapper(Function<TupleRowMapperBuilder, RowMapper<T4>> rowMapper) {
+        return mapper(rowMapper.apply(new TupleRowMapperBuilderImpl(getRowMapper)));
     }
 
 }