+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/mapper/MulitiQueryRowMapper5Impl.java
@@ -10,16 +10,16 @@ package cn.featherfly.hammer.sqldb.jdbc.mapper;
 
 import java.io.Serializable;
 import java.util.Map;
+import java.util.function.BiFunction;
+import java.util.function.Function;
 
 import org.apache.commons.lang3.ArrayUtils;
 
-import cn.featherfly.common.bean.InstantiatorFactory;
 import cn.featherfly.common.repository.mapper.MulitiQueryRowMapper5;
 import cn.featherfly.common.repository.mapper.MulitiQueryRowMapper6;
+import cn.featherfly.common.repository.mapper.PrefixedBeanMapper;
 import cn.featherfly.common.repository.mapper.RowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.Jdbc;
-import cn.featherfly.hammer.sqldb.jdbc.MapRowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.NestedBeanPropertyRowMapper;
+import cn.featherfly.common.repository.mapper.TupleRowMapperBuilder;
 
 /**
  * MulitiQueryRowMapper5.
@@ -38,36 +38,45 @@ public class MulitiQueryRowMapper5Impl<T1, T2, T3, T4, T5> extends AbstractMulit
      * Instantiates a new muliti query row mapper 5 impl.
      *
      * @param rowMappers the row mappers
-     * @param jdbc the jdbc
-     * @param instantiatorFactory the instantiator factory
+     * @param getRowMapper the get row mapper
      */
-    public MulitiQueryRowMapper5Impl(RowMapper<?>[] rowMappers, Jdbc jdbc, InstantiatorFactory instantiatorFactory) {
-        super(rowMappers, jdbc, instantiatorFactory);
+    public MulitiQueryRowMapper5Impl(RowMapper<?>[] rowMappers,
+        BiFunction<Class<?>, String, RowMapper<?>> getRowMapper) {
+        super(rowMappers, getRowMapper);
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public MulitiQueryRowMapper6<T1, T2, T3, T4, T5, Map<String, Serializable>> map() {
-        return map(new MapRowMapper(jdbc.getSqlTypeMappingManager()));
+    public MulitiQueryRowMapper6<T1, T2, T3, T4, T5, Map<String, Serializable>> mapper() {
+        return mapper(getRowMapper(PrefixedBeanMapper.MAP_TYPE));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T6> MulitiQueryRowMapper6<T1, T2, T3, T4, T5, T6> map(Class<T6> mappingType) {
-        return map(new NestedBeanPropertyRowMapper<>(instantiatorFactory.create(mappingType),
-            jdbc.getSqlTypeMappingManager()));
+    public <T6> MulitiQueryRowMapper6<T1, T2, T3, T4, T5, T6> mapper(Class<T6> mappingType) {
+        return mapper(getRowMapper(mappingType));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T6> MulitiQueryRowMapper6<T1, T2, T3, T4, T5, T6> map(RowMapper<T6> rowMapper) {
-        return new MulitiQueryRowMapper6Impl<>(ArrayUtils.add(getRowMappers(), rowMapper), jdbc, instantiatorFactory);
+    public <T6> MulitiQueryRowMapper6<T1, T2, T3, T4, T5, T6> mapper(RowMapper<T6> rowMapper) {
+        return new MulitiQueryRowMapper6Impl<>(ArrayUtils.add(getRowMappers(), rowMapper),
+            getRowMapper);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public <T6> MulitiQueryRowMapper6<T1, T2, T3, T4, T5, T6> mapper(
+        Function<TupleRowMapperBuilder, RowMapper<T6>> rowMapper) {
+        return mapper(rowMapper.apply(new TupleRowMapperBuilderImpl(getRowMapper)));
     }
 
 }