+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/mapper/MulitiQueryTupleMapperBuilderImpl.java
@@ -10,14 +10,14 @@ package cn.featherfly.hammer.sqldb.jdbc.mapper;
 
 import java.io.Serializable;
 import java.util.Map;
+import java.util.function.BiFunction;
+import java.util.function.Function;
 
-import cn.featherfly.common.bean.InstantiatorFactory;
 import cn.featherfly.common.repository.mapper.MulitiQueryRowMapper1;
 import cn.featherfly.common.repository.mapper.MulitiQueryTupleMapperBuilder;
+import cn.featherfly.common.repository.mapper.PrefixedBeanMapper;
 import cn.featherfly.common.repository.mapper.RowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.Jdbc;
-import cn.featherfly.hammer.sqldb.jdbc.MapRowMapper;
-import cn.featherfly.hammer.sqldb.jdbc.NestedBeanPropertyRowMapper;
+import cn.featherfly.common.repository.mapper.TupleRowMapperBuilder;
 
 /**
  * ArrayParamMulitiQueryTupleMapperBuilder.
@@ -26,45 +26,49 @@ import cn.featherfly.hammer.sqldb.jdbc.NestedBeanPropertyRowMapper;
  */
 public class MulitiQueryTupleMapperBuilderImpl implements MulitiQueryTupleMapperBuilder {
 
-    private final Jdbc jdbc;
-
-    private final InstantiatorFactory instantiatorFactory;
+    private final BiFunction<Class<?>, String, RowMapper<?>> getRowMapper;
 
     /**
      * Instantiates a new muliti query tuple mapper builder impl.
      *
-     * @param jdbc the jdbc
-     * @param instantiatorFactory the instantiator factory
+     * @param getRowMapper the get row mapper
      */
-    public MulitiQueryTupleMapperBuilderImpl(Jdbc jdbc, InstantiatorFactory instantiatorFactory) {
+    public MulitiQueryTupleMapperBuilderImpl(BiFunction<Class<?>, String, RowMapper<?>> getRowMapper) {
         super();
-        this.jdbc = jdbc;
-        this.instantiatorFactory = instantiatorFactory;
+        this.getRowMapper = getRowMapper;
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public MulitiQueryRowMapper1<Map<String, Serializable>> map() {
-        return map(new MapRowMapper(jdbc.getSqlTypeMappingManager()));
+    public MulitiQueryRowMapper1<Map<String, Serializable>> mapper() {
+        return mapper(PrefixedBeanMapper.MAP_TYPE);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T1> MulitiQueryRowMapper1<T1> mapper(Class<T1> mappingType) {
+        return mapper((RowMapper<T1>) getRowMapper.apply(mappingType, null));
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T1> MulitiQueryRowMapper1<T1> map(Class<T1> mappingType) {
-        return map(new NestedBeanPropertyRowMapper<>(instantiatorFactory.create(mappingType),
-            jdbc.getSqlTypeMappingManager()));
+    public <T1> MulitiQueryRowMapper1<T1> mapper(RowMapper<T1> rowMapper) {
+        return new MulitiQueryRowMapper1Impl<>(rowMapper, getRowMapper);
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public <T1> MulitiQueryRowMapper1<T1> map(RowMapper<T1> rowMapper) {
-        return new MulitiQueryRowMapper1Impl<>(rowMapper, jdbc, instantiatorFactory);
+    public <T1> MulitiQueryRowMapper1<T1> mapper(Function<TupleRowMapperBuilder, RowMapper<T1>> rowMapper) {
+        return mapper(rowMapper.apply(new TupleRowMapperBuilderImpl(getRowMapper)));
     }
 
 }