+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/mapper/TupleRowMapperBuilderImpl.java
@@ -11,7 +11,9 @@ package cn.featherfly.hammer.sqldb.jdbc.mapper;
 import java.util.function.BiFunction;
 
 import cn.featherfly.common.lang.Lang;
+import cn.featherfly.common.repository.mapper.PrefixedBeanMapper1;
 import cn.featherfly.common.repository.mapper.RowMapper;
+import cn.featherfly.common.repository.mapper.TupleRowMapperBuilder;
 import cn.featherfly.common.tuple.Tuple2;
 import cn.featherfly.common.tuple.Tuple3;
 import cn.featherfly.common.tuple.Tuple4;